requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
allure-pytest==2.13.2
allure-python-commons==2.13.2
annotated-types==0.6.0
anyio==4.3.0
astroid==3.1.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
black==24.3.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.8
easyimap==0.7.0
enum34==1.1.10
execnet==2.0.2
Faker==23.2.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
loguru==0.7.2
lxml==5.2.0
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
namedlist==1.8
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
playwright==1.41.2
pluggy==1.4.0
pycodestyle==2.11.1
pydantic==2.7.1
pydantic_core==2.18.2
pyee==11.0.1
pytest==8.0.1
pytest-base-url==2.1.0
pytest-mock==3.14.0
pytest-playwright==0.4.4
pytest-xdist==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
text-unidecode==1.3
tomlkit==0.12.4
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.29.0
yarl==1.9.4
